requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.6.0
cryptography==42.0.0
apscheduler==3.10.4
pytz==2024.1
playwright==1.41.0
python-dateutil==2.8.2
aiohttp==3.9.1
requests
python-docx==1.1.2
google-api-python-client==2.114.0
google-auth==2.27.0